+++ exporters/otlp/otlplog/otlploggrpc/internal/transform/log_test.go
@@ -30,73 +30,106 @@ var (
 	ts  = time.Date(2000, time.January, 0o1, 0, 0, 0, 0, time.FixedZone("GMT", 0))
 	obs = ts.Add(30 * time.Second)
 
+	tom   = api.String("user", "tom")
+	jerry = api.String("user", "jerry")
 	// A time before unix 0.
 	negativeTs = time.Date(1969, 7, 20, 20, 17, 0, 0, time.UTC)
 
-	alice = api.String("user", "alice")
-	bob   = api.String("user", "bob")
-
-	pbAlice = &cpb.KeyValue{Key: "user", Value: &cpb.AnyValue{
-		Value: &cpb.AnyValue_StringValue{StringValue: "alice"},
+	pbTom = &cpb.KeyValue{Key: "user", Value: &cpb.AnyValue{
+		Value: &cpb.AnyValue_StringValue{StringValue: "tom"},
 	}}
-	pbBob = &cpb.KeyValue{Key: "user", Value: &cpb.AnyValue{
-		Value: &cpb.AnyValue_StringValue{StringValue: "bob"},
+	pbJerry = &cpb.KeyValue{Key: "user", Value: &cpb.AnyValue{
+		Value: &cpb.AnyValue_StringValue{StringValue: "jerry"},
 	}}
 
-	sevA = api.SeverityInfo
-	sevB = api.SeverityError
+	sevC = api.SeverityInfo
+	sevD = api.SeverityError
 
-	pbSevA = lpb.SeverityNumber_SEVERITY_NUMBER_INFO
-	pbSevB = lpb.SeverityNumber_SEVERITY_NUMBER_ERROR
+	pbSevC = lpb.SeverityNumber_SEVERITY_NUMBER_INFO
+	pbSevD = lpb.SeverityNumber_SEVERITY_NUMBER_ERROR
 
-	bodyA = api.StringValue("a")
-	bodyB = api.StringValue("b")
+	bodyC = api.StringValue("c")
+	bodyD = api.StringValue("d")
 
-	pbBodyA = &cpb.AnyValue{
+	pbBodyC = &cpb.AnyValue{
 		Value: &cpb.AnyValue_StringValue{
-			StringValue: "a",
+			StringValue: "c",
 		},
 	}
-	pbBodyB = &cpb.AnyValue{
+	pbBodyD = &cpb.AnyValue{
 		Value: &cpb.AnyValue_StringValue{
-			StringValue: "b",
+			StringValue: "d",
 		},
 	}
 
-	spanIDA  = []byte{0, 0, 0, 0, 0, 0, 0, 1}
-	spanIDB  = []byte{0, 0, 0, 0, 0, 0, 0, 2}
-	traceIDA = []byte{0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1}
-	traceIDB = []byte{0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 2}
-	flagsA   = byte(1)
-	flagsB   = byte(0)
+	spanIDC  = []byte{0, 0, 0, 0, 0, 0, 0, 1}
+	spanIDD  = []byte{0, 0, 0, 0, 0, 0, 0, 2}
+	traceIDC = []byte{0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1}
+	traceIDD = []byte{0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 2}
+	flagsC   = byte(1)
+	flagsD   = byte(0)
 
 	scope = instrumentation.Scope{
-		Name:      "test/code/path",
-		Version:   "v0.1.0",
+		Name:      "otel/test/code/path1",
+		Version:   "v0.1.1",
 		SchemaURL: semconv.SchemaURL,
 	}
+	scope2 = instrumentation.Scope{
+		Name:      "otel/test/code/path2",
+		Version:   "v0.2.2",
+		SchemaURL: semconv.SchemaURL,
+	}
+	scopeList = []instrumentation.Scope{scope, scope2}
+
 	pbScope = &cpb.InstrumentationScope{
-		Name:    "test/code/path",
-		Version: "v0.1.0",
+		Name:    "otel/test/code/path1",
+		Version: "v0.1.1",
+	}
+	pbScope2 = &cpb.InstrumentationScope{
+		Name:    "otel/test/code/path2",
+		Version: "v0.2.2",
 	}
 
 	res = resource.NewWithAttributes(
 		semconv.SchemaURL,
-		semconv.ServiceName("test server"),
-		semconv.ServiceVersion("v0.1.0"),
+		semconv.ServiceName("service1"),
+		semconv.ServiceVersion("v0.1.1"),
 	)
+	res2 = resource.NewWithAttributes(
+		semconv.SchemaURL,
+		semconv.ServiceName("service2"),
+		semconv.ServiceVersion("v0.2.2"),
+	)
+	resList = []*resource.Resource{res, res2}
+
 	pbRes = &rpb.Resource{
 		Attributes: []*cpb.KeyValue{
 			{
 				Key: "service.name",
 				Value: &cpb.AnyValue{
-					Value: &cpb.AnyValue_StringValue{StringValue: "test server"},
+					Value: &cpb.AnyValue_StringValue{StringValue: "service1"},
 				},
 			},
 			{
 				Key: "service.version",
 				Value: &cpb.AnyValue{
-					Value: &cpb.AnyValue_StringValue{StringValue: "v0.1.0"},
+					Value: &cpb.AnyValue_StringValue{StringValue: "v0.1.1"},
+				},
+			},
+		},
+	}
+	pbRes2 = &rpb.Resource{
+		Attributes: []*cpb.KeyValue{
+			{
+				Key: "service.name",
+				Value: &cpb.AnyValue{
+					Value: &cpb.AnyValue_StringValue{StringValue: "service2"},
+				},
+			},
+			{
+				Key: "service.version",
+				Value: &cpb.AnyValue{
+					Value: &cpb.AnyValue_StringValue{StringValue: "v0.2.2"},
 				},
 			},
 		},
@@ -105,75 +138,79 @@ var (
 	records = func() []log.Record {
 		var out []log.Record
 
-		out = append(out, logtest.RecordFactory{
-			Timestamp:            ts,
-			ObservedTimestamp:    obs,
-			Severity:             sevA,
-			SeverityText:         "A",
-			Body:                 bodyA,
-			Attributes:           []api.KeyValue{alice},
-			TraceID:              trace.TraceID(traceIDA),
-			SpanID:               trace.SpanID(spanIDA),
-			TraceFlags:           trace.TraceFlags(flagsA),
-			InstrumentationScope: &scope,
-			Resource:             res,
-		}.NewRecord())
+		for _, r := range resList {
+			for _, s := range scopeList {
+				out = append(out, logtest.RecordFactory{
+					Timestamp:            ts,
+					ObservedTimestamp:    obs,
+					Severity:             sevC,
+					SeverityText:         "C",
+					Body:                 bodyC,
+					Attributes:           []api.KeyValue{tom},
+					TraceID:              trace.TraceID(traceIDC),
+					SpanID:               trace.SpanID(spanIDC),
+					TraceFlags:           trace.TraceFlags(flagsC),
+					InstrumentationScope: &s,
+					Resource:             r,
+				}.NewRecord())
 
-		out = append(out, logtest.RecordFactory{
-			Timestamp:            ts,
-			ObservedTimestamp:    obs,
-			Severity:             sevA,
-			SeverityText:         "A",
-			Body:                 bodyA,
-			Attributes:           []api.KeyValue{bob},
-			TraceID:              trace.TraceID(traceIDA),
-			SpanID:               trace.SpanID(spanIDA),
-			TraceFlags:           trace.TraceFlags(flagsA),
-			InstrumentationScope: &scope,
-			Resource:             res,
-		}.NewRecord())
+				out = append(out, logtest.RecordFactory{
+					Timestamp:            ts,
+					ObservedTimestamp:    obs,
+					Severity:             sevC,
+					SeverityText:         "C",
+					Body:                 bodyC,
+					Attributes:           []api.KeyValue{jerry},
+					TraceID:              trace.TraceID(traceIDC),
+					SpanID:               trace.SpanID(spanIDC),
+					TraceFlags:           trace.TraceFlags(flagsC),
+					InstrumentationScope: &s,
+					Resource:             r,
+				}.NewRecord())
 
-		out = append(out, logtest.RecordFactory{
-			Timestamp:            ts,
-			ObservedTimestamp:    obs,
-			Severity:             sevB,
-			SeverityText:         "B",
-			Body:                 bodyB,
-			Attributes:           []api.KeyValue{alice},
-			TraceID:              trace.TraceID(traceIDB),
-			SpanID:               trace.SpanID(spanIDB),
-			TraceFlags:           trace.TraceFlags(flagsB),
-			InstrumentationScope: &scope,
-			Resource:             res,
-		}.NewRecord())
+				out = append(out, logtest.RecordFactory{
+					Timestamp:            ts,
+					ObservedTimestamp:    obs,
+					Severity:             sevD,
+					SeverityText:         "D",
+					Body:                 bodyD,
+					Attributes:           []api.KeyValue{tom},
+					TraceID:              trace.TraceID(traceIDD),
+					SpanID:               trace.SpanID(spanIDD),
+					TraceFlags:           trace.TraceFlags(flagsD),
+					InstrumentationScope: &s,
+					Resource:             r,
+				}.NewRecord())
 
-		out = append(out, logtest.RecordFactory{
-			Timestamp:            ts,
-			ObservedTimestamp:    obs,
-			Severity:             sevB,
-			SeverityText:         "B",
-			Body:                 bodyB,
-			Attributes:           []api.KeyValue{bob},
-			TraceID:              trace.TraceID(traceIDB),
-			SpanID:               trace.SpanID(spanIDB),
-			TraceFlags:           trace.TraceFlags(flagsB),
-			InstrumentationScope: &scope,
-			Resource:             res,
-		}.NewRecord())
+				out = append(out, logtest.RecordFactory{
+					Timestamp:            ts,
+					ObservedTimestamp:    obs,
+					Severity:             sevD,
+					SeverityText:         "D",
+					Body:                 bodyD,
+					Attributes:           []api.KeyValue{jerry},
+					TraceID:              trace.TraceID(traceIDD),
+					SpanID:               trace.SpanID(spanIDD),
+					TraceFlags:           trace.TraceFlags(flagsD),
+					InstrumentationScope: &s,
+					Resource:             r,
+				}.NewRecord())
 
-		out = append(out, logtest.RecordFactory{
-			Timestamp:            negativeTs,
-			ObservedTimestamp:    obs,
-			Severity:             sevB,
-			SeverityText:         "B",
-			Body:                 bodyB,
-			Attributes:           []api.KeyValue{bob},
-			TraceID:              trace.TraceID(traceIDB),
-			SpanID:               trace.SpanID(spanIDB),
-			TraceFlags:           trace.TraceFlags(flagsB),
-			InstrumentationScope: &scope,
-			Resource:             res,
-		}.NewRecord())
+				out = append(out, logtest.RecordFactory{
+					Timestamp:            negativeTs,
+					ObservedTimestamp:    obs,
+					Severity:             sevD,
+					SeverityText:         "D",
+					Body:                 bodyD,
+					Attributes:           []api.KeyValue{jerry},
+					TraceID:              trace.TraceID(traceIDD),
+					SpanID:               trace.SpanID(spanIDD),
+					TraceFlags:           trace.TraceFlags(flagsD),
+					InstrumentationScope: &s,
+					Resource:             r,
+				}.NewRecord())
+			}
+		}
 
 		return out
 	}()
@@ -182,76 +219,90 @@ var (
 		{
 			TimeUnixNano:         uint64(ts.UnixNano()),
 			ObservedTimeUnixNano: uint64(obs.UnixNano()),
-			SeverityNumber:       pbSevA,
-			SeverityText:         "A",
-			Body:                 pbBodyA,
-			Attributes:           []*cpb.KeyValue{pbAlice},
-			Flags:                uint32(flagsA),
-			TraceId:              traceIDA,
-			SpanId:               spanIDA,
+			SeverityNumber:       pbSevC,
+			SeverityText:         "C",
+			Body:                 pbBodyC,
+			Attributes:           []*cpb.KeyValue{pbTom},
+			Flags:                uint32(flagsC),
+			TraceId:              traceIDC,
+			SpanId:               spanIDC,
 		},
 		{
 			TimeUnixNano:         uint64(ts.UnixNano()),
 			ObservedTimeUnixNano: uint64(obs.UnixNano()),
-			SeverityNumber:       pbSevA,
-			SeverityText:         "A",
-			Body:                 pbBodyA,
-			Attributes:           []*cpb.KeyValue{pbBob},
-			Flags:                uint32(flagsA),
-			TraceId:              traceIDA,
-			SpanId:               spanIDA,
+			SeverityNumber:       pbSevC,
+			SeverityText:         "C",
+			Body:                 pbBodyC,
+			Attributes:           []*cpb.KeyValue{pbJerry},
+			Flags:                uint32(flagsC),
+			TraceId:              traceIDC,
+			SpanId:               spanIDC,
 		},
 		{
 			TimeUnixNano:         uint64(ts.UnixNano()),
 			ObservedTimeUnixNano: uint64(obs.UnixNano()),
-			SeverityNumber:       pbSevB,
-			SeverityText:         "B",
-			Body:                 pbBodyB,
-			Attributes:           []*cpb.KeyValue{pbAlice},
-			Flags:                uint32(flagsB),
-			TraceId:              traceIDB,
-			SpanId:               spanIDB,
+			SeverityNumber:       pbSevD,
+			SeverityText:         "D",
+			Body:                 pbBodyD,
+			Attributes:           []*cpb.KeyValue{pbTom},
+			Flags:                uint32(flagsD),
+			TraceId:              traceIDD,
+			SpanId:               spanIDD,
 		},
 		{
 			TimeUnixNano:         uint64(ts.UnixNano()),
 			ObservedTimeUnixNano: uint64(obs.UnixNano()),
-			SeverityNumber:       pbSevB,
-			SeverityText:         "B",
-			Body:                 pbBodyB,
-			Attributes:           []*cpb.KeyValue{pbBob},
-			Flags:                uint32(flagsB),
-			TraceId:              traceIDB,
-			SpanId:               spanIDB,
+			SeverityNumber:       pbSevD,
+			SeverityText:         "D",
+			Body:                 pbBodyD,
+			Attributes:           []*cpb.KeyValue{pbJerry},
+			Flags:                uint32(flagsD),
+			TraceId:              traceIDD,
+			SpanId:               spanIDD,
 		},
 		{
 			TimeUnixNano:         0,
 			ObservedTimeUnixNano: uint64(obs.UnixNano()),
-			SeverityNumber:       pbSevB,
-			SeverityText:         "B",
-			Body:                 pbBodyB,
-			Attributes:           []*cpb.KeyValue{pbBob},
-			Flags:                uint32(flagsB),
-			TraceId:              traceIDB,
-			SpanId:               spanIDB,
+			SeverityNumber:       pbSevD,
+			SeverityText:         "D",
+			Body:                 pbBodyD,
+			Attributes:           []*cpb.KeyValue{pbJerry},
+			Flags:                uint32(flagsD),
+			TraceId:              traceIDD,
+			SpanId:               spanIDD,
 		},
 	}
 
-	pbScopeLogs = &lpb.ScopeLogs{
-		Scope:      pbScope,
-		SchemaUrl:  semconv.SchemaURL,
-		LogRecords: pbLogRecords,
+	pbScopeLogsList = []*lpb.ScopeLogs{
+		{
+			Scope:      pbScope,
+			SchemaUrl:  semconv.SchemaURL,
+			LogRecords: pbLogRecords,
+		},
+		{
+			Scope:      pbScope2,
+			SchemaUrl:  semconv.SchemaURL,
+			LogRecords: pbLogRecords,
+		},
 	}
 
-	pbResourceLogs = &lpb.ResourceLogs{
-		Resource:  pbRes,
-		SchemaUrl: semconv.SchemaURL,
-		ScopeLogs: []*lpb.ScopeLogs{pbScopeLogs},
+	pbResourceLogsList = []*lpb.ResourceLogs{
+		{
+			Resource:  pbRes,
+			SchemaUrl: semconv.SchemaURL,
+			ScopeLogs: pbScopeLogsList,
+		},
+		{
+			Resource:  pbRes2,
+			SchemaUrl: semconv.SchemaURL,
+			ScopeLogs: pbScopeLogsList,
+		},
 	}
 )
 
 func TestResourceLogs(t *testing.T) {
-	want := []*lpb.ResourceLogs{pbResourceLogs}
-	assert.Equal(t, want, ResourceLogs(records))
+	want := pbResourceLogsList
+	assert.ElementsMatch(t, want, ResourceLogs(records))
 }
 
 func TestSeverityNumber(t *testing.T) {